more refining
