+++ src/Custom/Assistants/AssistantClient.Protocol.cs
@@ -257,7 +257,7 @@ public partial class AssistantClient
 
     public virtual IEnumerable<ClientResult> GetMessages(string threadId, int? limit, string order, string after, string before, RequestOptions options)
     {
-        IEnumerator<ClientResult> enumerator = new MessagePageResultEnumerator(_messageSubClient, threadId, limit, order, after, before, options);
+        IEnumerator<ClientResult> enumerator = new MessageCollectionClient(_messageSubClient, threadId, limit, order, after, before, options);
         return PageCollectionHelpers.CreateProtocol(enumerator);
     }
