--- src/To.Be.Generated/Internal/PageToken.cs
@@ -1,13 +1,0 @@
-﻿//using System;
-//using System.ClientModel;
-//using System.Collections.Generic;
-//using System.Text;
-
-//#nullable enable
-
-//namespace OpenAI;
-
-//internal abstract class PageToken : ContinuationToken
-//{
-//    public abstract PageToken? GetNextPageToken();
-//}
